requirements.txt


#################

streamlit==1.35.0
PyPDF2==3.0.1
langchain
langchain-core
langchain-google-genai
google-generativeai==0.4.1
faiss-cpu
python-dotenv==1.0.1






